STING/profile_service/requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.2
psycopg2-binary>=2.9.0
hvac>=1.2.1
requests>=2.31.0
Pillow>=10.0.0
python-magic>=0.4.27
cryptography>=41.0.0
pydantic>=2.0.0
gunicorn==20.1.0